requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
django>=2.1.5
django-allauth==0.38.0
django-heroku==0.3.1
django-widget-tweaks==1.4.3
gunicorn==19.9.0
idna==2.8
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
urllib3==1.24.1
whitenoise==4.1.2
